7 store(e-commerce website)/requirements.txt
Flask==2.2.5
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.1.1
email-validator==2.0.0
python-dotenv==1.0.0
stripe==5.5.0
Werkzeug==2.2.3
Flask-Login==0.6.2
